requirements.txt
enum34
fusepy
kaitaistruct==0.7
